requirements.txt
cachetools==5.5.0
certifi
charset-normalizer==3.4.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.67.1
grpcio-status==1.67.1
idna==3.10
packaging==24.2
proto-plus==1.25.0
protobuf
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.3